tests: Use PRIuPTR and uintptr_t for formatting size_t
In particular, remove the non-portable assumption that
sizeof(long) >= sizeof(size_t) in formatted output.

(cherry picked from commit c63fb3c33a80cf85c6ed2ad52e2fbd1b51f672cf)

+++ tests/test.h
@@ -252,7 +252,7 @@ static char test_sections[255];
 static void *_test_malloc(size_t sz, const char *curloc) {
   void *ptr;
   ptr = malloc(sz);
-  if (ptr == NULL) FATALERR("Failed to malloc(%lu) bytes at %s\n",(unsigned long)sz,curloc);
+  if (ptr == NULL) FATALERR("Failed to malloc(%" PRIuPTR ") bytes at %s\n",(uintptr_t)sz,curloc);
   return ptr;
 }
 static void *_test_calloc(size_t sz, const char *curloc) {
